Rocketry - Learning how to launch Rockets with Habitat.
This is a small sample web application written in Rust using the Rocket
project and includes a Habitat plan which can build this into a package.
It takes one optional command line argument which is a port number to
use, otherwise defaulting to `8000` so that it can run without root
permissions. The simple Rocket application is setup to run in
"production" mode and will skip all `Rocket.toml` and environment
variables. This isn't a requirement per se, but illustrative of how
Habitat can manage these details dynamically on behalf of a service.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>